Move split to utils file.
