+++ stwo_cairo_prover/crates/prover/src/components/range_check_builtin/simd_trace.rs
@@ -1,11 +1,10 @@
-use std::ops::{BitAnd, BitOrAssign, Shl, ShrAssign};
 use std::simd::u32x16;
 
 use itertools::{chain, Itertools};
 use num_traits::Zero;
 use stwo_prover::constraint_framework::logup::LogupTraceGenerator;
 use stwo_prover::core::backend::simd::column::BaseColumn;
-use stwo_prover::core::backend::simd::m31::{PackedM31, LOG_N_LANES, N_LANES};
+use stwo_prover::core::backend::simd::m31::{PackedM31, LOG_N_LANES};
 use stwo_prover::core::backend::simd::qm31::PackedQM31;
 use stwo_prover::core::backend::simd::SimdBackend;
 use stwo_prover::core::backend::{Col, Column};
@@ -16,15 +15,16 @@ use stwo_prover::core::poly::BitReversedOrder;
 use stwo_prover::core::ColumnVec;
 
 use super::component::{LAST_VALUE_OFFSET, N_VALUES_FELTS};
-use crate::components::memory::component::N_M31_IN_FELT252;
-use crate::components::memory::{MemoryLookupElements, N_ADDRESS_FELTS, N_BITS_PER_FELT};
+use crate::components::memory::{MemoryLookupElements, N_ADDRESS_FELTS};
 use crate::components::range_check_unit::RangeElements;
+use crate::felt::split_u128_simd;
 
 // Memory addresses and the corresponding values, for the RangeCheck128Builtin segment.
 pub struct RangeCheckInput {
     pub address: PackedM31,
     pub values: [u32x16; 4],
 }
+
 impl Default for RangeCheckInput {
     fn default() -> Self {
         Self {
@@ -112,69 +112,6 @@ pub fn gen_interaction_trace(
     logup_gen.finalize()
 }
 
-/// Splits a 32N bit dense representation into felts, each with N_BITS_PER_FELT bits.
-///
-/// Parameters:
-/// - `N`: the number of 32-bit words in the input.
-/// - `M`: the number of felts in the output.
-/// - `TU32`: the type of the input and output words.
-/// - `x`: the input dense representation.
-/// - `mask`: (1 << N_BITS_PER_FELT) - 1.
-fn split<const N: usize, const M: usize, TU32>(x: [TU32; N], mask: TU32) -> [TU32; M]
-where
-    TU32: BitAnd<Output = TU32>
-        + BitOrAssign
-        + Copy
-        + ShrAssign<u32>
-        + Shl<u32, Output = TU32>
-        + Default,
-{
-    let mut res = [TU32::default(); M];
-    let mut n_bits_in_word = 32;
-    let mut word_i = 0;
-    let mut word = x[word_i];
-    for e in res.iter_mut() {
-        // If current word has more bits than needed, chop it.
-        if n_bits_in_word > N_BITS_PER_FELT {
-            *e = word & mask;
-            word >>= N_BITS_PER_FELT as u32;
-            n_bits_in_word -= N_BITS_PER_FELT;
-            continue;
-        }
-
-        *e = word;
-        // Fetch next word.
-        word_i += 1;
-        word = x.get(word_i).copied().unwrap_or_default();
-
-        // If we need more bits to fill, take from next word.
-        if n_bits_in_word < N_BITS_PER_FELT {
-            *e |= (word << n_bits_in_word as u32) & mask;
-            word >>= (N_BITS_PER_FELT - n_bits_in_word) as u32;
-        }
-
-        n_bits_in_word += 32 - N_BITS_PER_FELT;
-    }
-    res
-}
-
-/// Splits a 128 bit dense representation into felts, each with N_BITS_PER_FELT bits.
-pub fn split_u128_simd(x: [u32x16; 4]) -> [PackedM31; N_VALUES_FELTS] {
-    split(x, u32x16::from_array([(1 << N_BITS_PER_FELT) - 1; N_LANES]))
-        .map(|x| PackedM31::from(x.to_array().map(M31::from_u32_unchecked)))
-}
-
-/// Splits a 252 bit dense representation into felts, each with N_BITS_PER_FELT bits.
-pub fn split_f252_simd(x: [u32x16; 8]) -> [PackedM31; N_M31_IN_FELT252] {
-    split(x, u32x16::from_array([(1 << N_BITS_PER_FELT) - 1; N_LANES]))
-        .map(|x| PackedM31::from(x.to_array().map(M31::from_u32_unchecked)))
-}
-
-/// Splits a 252 bit dense representation into felts, each with N_BITS_PER_FELT bits.
-pub fn split_f252(x: [u32; 8]) -> [M31; N_M31_IN_FELT252] {
-    split(x, (1 << N_BITS_PER_FELT) - 1).map(M31::from_u32_unchecked)
-}
-
 #[cfg(test)]
 mod tests {
     use std::array;
@@ -183,14 +120,16 @@ mod tests {
     use rand::Rng;
     use stwo_prover::constraint_framework::constant_columns::gen_is_first;
     use stwo_prover::constraint_framework::FrameworkComponent;
+    use stwo_prover::core::backend::simd::m31::N_LANES;
     use stwo_prover::core::channel::{Blake2sChannel, Channel};
     use stwo_prover::core::fields::IntoSlice;
     use stwo_prover::core::pcs::TreeVec;
     use stwo_prover::core::vcs::blake2_hash::Blake2sHasher;
 
     use super::*;
-    use crate::components::memory::MemoryLookupElements;
+    use crate::components::memory::{MemoryLookupElements, N_BITS_PER_FELT};
     use crate::components::range_check_builtin::component::RangeCheck128BuiltinComponent;
+    use crate::felt::split_f252;
 
     #[test]
     fn test_generate_trace() {
