+++ stwo_cairo_prover/crates/prover/src/lib.rs
@@ -1,6 +1,7 @@
 #![feature(array_methods, portable_simd, iter_array_chunks)]
 pub mod cairo_air;
 pub mod components;
+pub mod felt;
 pub mod input;
 pub mod prover_types;
